Rename call store to 'heap call store'
