+++ src/HeapCallStore.js
@@ -1,0 +1,17 @@
+module.exports = HeapCallStore;
+
+function HeapCallStore() {
+	var calls = [];
+
+	this.record = function record(thisBinding, callArguments, onExecuted) {
+		calls.push({
+			thisBinding: thisBinding, 
+			callArguments : callArguments,
+			onExecuted : onExecuted
+		});
+	};
+
+	this.getCalls = function getCalls() {
+		return calls.slice(0); // clone
+	};
+}
